+++ test/settings/components/RealTimeUpdatesSettings.test.tsx
@@ -13,7 +13,7 @@ describe('<RealTimeUpdatesSettings />', () => {
     <SettingsProvider value={fromPartial({ realTimeUpdates })}>
       <RealTimeUpdatesSettings
         toggleRealTimeUpdates={toggleRealTimeUpdates}
-        setRealTimeUpdatesInterval={setRealTimeUpdatesInterval}
+        onIntervalChange={setRealTimeUpdatesInterval}
       />
     </SettingsProvider>,
   );